requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.1
colorlog==6.7.0
frozenlist==1.4.0
hikari==2.0.0.dev121
hikari-lightbulb==2.3.3
hikari-miru==3.2.2
idna==3.4
lavasnek-rs==0.1.0a5
multidict==6.0.4
python-dotenv==1.0.0
uvloop==0.19.0
yarl==1.9.2
